+++ packages/std/vitest.config.ts
@@ -9,6 +9,23 @@ export default defineConfig({
     },
   },
   test: {
+    coverage: {
+      all: true,
+      include: ["src/**/*.ts"],
+      exclude: ["src/**/*.test.ts", "src/index.ts"],
+      reporter: ["text", "json", "html", "lcov"],
+      reportsDirectory: "coverage",
+      thresholds: {
+        lines: 94,
+        functions: 96,
+        branches: 90,
+        statements: 94,
+      },
+    },
     include: ["src/**/*.test.ts"],
+    typecheck: {
+      enabled: true,
+      include: ["src/**/*.test.ts"],
+    },
   },
 });